requirements.txt
pyusb==1.1.1
requests==2.25.1
Flask==1.1.2
Flask-RESTful==0.3.9
